requirements.txt
torch
torchvision
transformers
numpy
pandas
faiss-cpu
opencv-python
streamlit
scikit-learn
PILLOW
